+++ integration/flink/flink2/src/main/java/io/openlineage/flink/visitor/identifier/KafkaTableLineageDatasetIdentifierVisitor.java
@@ -6,16 +6,18 @@
 package io.openlineage.flink.visitor.identifier;
 
 import io.openlineage.client.utils.DatasetIdentifier;
-import io.openlineage.client.utils.DatasetIdentifier.Symlink;
-import io.openlineage.client.utils.DatasetIdentifier.SymlinkType;
+import io.openlineage.flink.visitor.identifier.catalog.CatalogSymlinkProvider;
+import io.openlineage.flink.visitor.identifier.catalog.GenericInMemoryCatalogSymlinkProvider;
 import io.openlineage.flink.wrapper.TableLineageDatasetWrapper;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.flink.streaming.api.lineage.LineageDataset;
 import org.apache.flink.table.catalog.CatalogBaseTable;
+import org.apache.flink.table.catalog.listener.CatalogContext;
 import org.apache.flink.table.planner.lineage.TableLineageDataset;
 
 /** Class to extract dataset identifier from {@link TableLineageDataset} stored on Kafka. */
@@ -25,6 +27,12 @@ public class KafkaTableLineageDatasetIdentifierVisitor implements DatasetIdentif
   private static final String KAFKA_DATASET_PREFIX = "kafka://";
   private static final String COMMA = ",";
   private static final String SEMICOLON = ";";
+
+  private final List<CatalogSymlinkProvider> catalogSymlinkProviders;
+
+  public KafkaTableLineageDatasetIdentifierVisitor() {
+    catalogSymlinkProviders = List.of(new GenericInMemoryCatalogSymlinkProvider());
+  }
 
   @Override
   public boolean isDefinedAt(LineageDataset dataset) {
@@ -36,8 +44,6 @@ public class KafkaTableLineageDatasetIdentifierVisitor implements DatasetIdentif
       log.info("Table is null for dataset {}", dataset);
       return false;
     }
-
-    // TODO: get comment from catalogBaseTable's comment field
 
     Map<String, String> options = table.getOptions();
     if (options == null) {
@@ -56,7 +62,9 @@ public class KafkaTableLineageDatasetIdentifierVisitor implements DatasetIdentif
 
   @Override
   public Collection<DatasetIdentifier> apply(LineageDataset dataset) {
-    CatalogBaseTable table = new TableLineageDatasetWrapper(dataset).getTable().orElseThrow();
+    TableLineageDatasetWrapper wrapper = new TableLineageDatasetWrapper(dataset);
+    CatalogBaseTable table = wrapper.getTable().orElseThrow();
+    Optional<CatalogContext> catalogContext = wrapper.getCatalogContext();
 
     if (log.isDebugEnabled()) {
       log.debug(
@@ -65,15 +73,26 @@ public class KafkaTableLineageDatasetIdentifierVisitor implements DatasetIdentif
           table.getOptions().get("topic"));
     }
 
-    return Collections.singletonList(
+    DatasetIdentifier identifier =
         datasetIdentifierForKafka(
             table.getOptions().get("properties.bootstrap.servers"),
-            table.getOptions().get("topic"),
-            dataset.name()));
+            table.getOptions().get("topic"));
+
+    catalogContext
+        .flatMap(
+            context ->
+                catalogSymlinkProviders.stream()
+                    .filter(c -> c.isDefinedAt(context.getClazz()))
+                    .map(p -> p.getSymlink(context, dataset))
+                    .filter(Optional::isPresent)
+                    .map(Optional::get)
+                    .findFirst())
+        .ifPresent(identifier::withSymlink);
+
+    return Collections.singletonList(identifier);
   }
 
-  private DatasetIdentifier datasetIdentifierForKafka(
-      String bootstrapServers, String topic, String tableName) {
+  private DatasetIdentifier datasetIdentifierForKafka(String bootstrapServers, String topic) {
     String kafkaHost = bootstrapServers;
     if (bootstrapServers.contains(COMMA)) {
       kafkaHost = bootstrapServers.split(COMMA)[0];
@@ -83,7 +102,6 @@ public class KafkaTableLineageDatasetIdentifierVisitor implements DatasetIdentif
 
     String namespace = String.format(KAFKA_DATASET_PREFIX + kafkaHost);
 
-    return new DatasetIdentifier(
-        topic, namespace, List.of(new Symlink(tableName, namespace, SymlinkType.TABLE)));
+    return new DatasetIdentifier(topic, namespace);
   }
 }